Separate into config files (#129)
