+++ runtime/devnet/src/lib.rs
@@ -14,19 +14,18 @@ use gafi_primitives::membership::{Achievements, MembershipLevelPoints};
 use pallet_grandpa::{
 	fg_primitives, AuthorityId as GrandpaId, AuthorityList as GrandpaAuthorityList,
 };
-use runtime_common::{prod_or_fast, impls::DealWithFees};
+use runtime_common::{impls::DealWithFees, prod_or_fast};
 use sp_api::impl_runtime_apis;
 use sp_consensus_aura::sr25519::AuthorityId as AuraId;
 use sp_core::{
 	crypto::{ByteArray, KeyTypeId},
 	OpaqueMetadata, H160, H256, U256,
 };
-use sp_io::hashing::blake2_256;
 use sp_runtime::{
 	create_runtime_str, generic, impl_opaque_keys,
 	traits::{
-		AccountIdLookup, BlakeTwo256, Block as BlockT, DispatchInfoOf, Dispatchable,
-		IdentifyAccount, NumberFor, PostDispatchInfoOf, Verify,
+		BlakeTwo256, Block as BlockT, DispatchInfoOf, Dispatchable, IdentifyAccount, NumberFor,
+		PostDispatchInfoOf, Verify,
 	},
 	transaction_validity::{TransactionSource, TransactionValidity, TransactionValidityError},
 	ApplyExtrinsicResult, MultiSignature,
@@ -35,7 +34,6 @@ use sp_std::{cmp::Ordering, marker::PhantomData, prelude::*};
 #[cfg(feature = "std")]
 use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
-use static_assertions::const_assert;
 
 // A few exports that help ease life for downstream crates.
 use fp_rpc::TransactionStatus;
@@ -55,15 +53,14 @@ pub use frame_system::{Call as SystemCall, EnsureRoot};
 pub use pallet_balances::Call as BalancesCall;
 use pallet_ethereum::{Call::transact, Transaction as EthereumTransaction};
 use pallet_evm::{
-	Account as EVMAccount, EVMCurrencyAdapter, EnsureAddressNever, EnsureAddressRoot,
-	FeeCalculator, HashedAddressMapping, Runner,
+	Account as EVMAccount, EVMCurrencyAdapter, FeeCalculator, HashedAddressMapping, Runner,
 };
 pub use pallet_timestamp::Call as TimestampCall;
 pub use pallet_transaction_payment::CurrencyAdapter;
 #[cfg(any(feature = "std", test))]
 pub use sp_runtime::BuildStorage;
-pub use sp_runtime::{Perbill, Permill};
 use sp_runtime::SaturatedConversion;
+pub use sp_runtime::{Perbill, Permill};
 
 pub use gafi_primitives::{
 	cache::Cache,
@@ -85,16 +82,15 @@ pub use pallet_faucet;
 pub use pallet_player;
 pub use pallet_pool;
 pub use pallet_pool_names;
+pub use pallet_whitelist;
 pub use sponsored_pool;
 pub use staking_pool;
 pub use upfront_pool;
-pub use pallet_whitelist;
 
-// custom traits
-use gafi_tx::{GafiEVMCurrencyAdapter, GafiGasWeightMapping};
+mod pallets;
+use pallets::*;
 
 mod precompiles;
-use precompiles::FrontierPrecompiles;
 
 /// Type of block number.
 pub type BlockNumber = u32;
@@ -195,58 +191,6 @@ parameter_types! {
 
 // Configure FRAME pallets to include in runtime.
 
-impl frame_system::Config for Runtime {
-	/// The basic call filter to use in dispatchable.
-	type BaseCallFilter = frame_support::traits::Everything;
-	/// Block & extrinsics weights: base values and limits.
-	type BlockWeights = BlockWeights;
-	/// The maximum length of a block (in bytes).
-	type BlockLength = BlockLength;
-	/// The identifier used to distinguish between accounts.
-	type AccountId = AccountId;
-	/// The aggregated dispatch type that is available for extrinsics.
-	type Call = Call;
-	/// The lookup mechanism to get account ID from whatever is passed in dispatchers.
-	type Lookup = AccountIdLookup<AccountId, ()>;
-	/// The index type for storing how many extrinsics an account has signed.
-	type Index = Index;
-	/// The index type for blocks.
-	type BlockNumber = BlockNumber;
-	/// The type for hashing blocks and tries.
-	type Hash = Hash;
-	/// The hashing algorithm used.
-	type Hashing = BlakeTwo256;
-	/// The header type.
-	type Header = generic::Header<BlockNumber, BlakeTwo256>;
-	/// The ubiquitous event type.
-	type Event = Event;
-	/// The ubiquitous origin type.
-	type Origin = Origin;
-	/// Maximum number of block number to block hash mappings to keep (oldest pruned first).
-	type BlockHashCount = BlockHashCount;
-	/// The weight of database operations that the runtime can invoke.
-	type DbWeight = RocksDbWeight;
-	/// Version of the runtime.
-	type Version = Version;
-	/// Converts a module to the index of the module in `construct_runtime!`.
-	///
-	/// This type is being generated by `construct_runtime!`.
-	type PalletInfo = PalletInfo;
-	/// What to do if a new account is created.
-	type OnNewAccount = ();
-	/// What to do if an account is fully reaped from the system.
-	type OnKilledAccount = ();
-	/// The data to be stored in an account.
-	type AccountData = pallet_balances::AccountData<Balance>;
-	/// Weight information for the extrinsics of this pallet.
-	type SystemWeightInfo = ();
-	/// This is used as an identifier of the chain. 42 is the generic substrate prefix.
-	type SS58Prefix = SS58Prefix;
-	/// The set code logic, just the default since we're not a parachain.
-	type OnSetCode = ();
-	type MaxConsumers = ConstU32<16>;
-}
-
 parameter_types! {
 	pub const MaxAuthorities: u32 = 100;
 }
@@ -255,71 +199,6 @@ impl pallet_aura::Config for Runtime {
 	type AuthorityId = AuraId;
 	type DisabledValidators = ();
 	type MaxAuthorities = MaxAuthorities;
-}
-
-impl pallet_grandpa::Config for Runtime {
-	type Event = Event;
-	type Call = Call;
-
-	type KeyOwnerProofSystem = ();
-
-	type KeyOwnerProof =
-		<Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(KeyTypeId, GrandpaId)>>::Proof;
-
-	type KeyOwnerIdentification = <Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(
-		KeyTypeId,
-		GrandpaId,
-	)>>::IdentificationTuple;
-
-	type HandleEquivocation = ();
-
-	type WeightInfo = ();
-	type MaxAuthorities = ConstU32<32>;
-}
-
-parameter_types! {
-	pub const MinimumPeriod: u64 = SLOT_DURATION / 2;
-}
-
-impl pallet_timestamp::Config for Runtime {
-	/// A timestamp: milliseconds since the unix epoch.
-	type Moment = u64;
-	type MinimumPeriod = MinimumPeriod;
-	type WeightInfo = ();
-	#[cfg(all(feature = "aura", not(feature = "manual-seal")))]
-	type OnTimestampSet = Aura;
-	#[cfg(feature = "manual-seal")]
-	type OnTimestampSet = ();
-}
-
-parameter_types! {
-	pub  NativeTokenExistentialDeposit: Balance = 1 * unit(GAKI); // 1 GAKI
-	// For weight estimation, we assume that the most locks on an individual account will be 50.
-	// This number may need to be adjusted in the future if this assumption no longer holds true.
-	pub const MaxLocks: u32 = 50;
-}
-
-impl pallet_balances::Config for Runtime {
-	type MaxLocks = MaxLocks;
-	type MaxReserves = ();
-	type ReserveIdentifier = [u8; 8];
-	/// The type for recording an account's balance.
-	type Balance = Balance;
-	/// The ubiquitous event type.
-	type Event = Event;
-	type DustRemoval = ();
-	type ExistentialDeposit = NativeTokenExistentialDeposit;
-	type AccountStore = System;
-	type WeightInfo = ();
-}
-
-impl pallet_transaction_payment::Config for Runtime {
-	type Event = Event;
-	type OnChargeTransaction = CurrencyAdapter<Balances, DealWithFees<Runtime>>;
-	type OperationalFeeMultiplier = ConstU8<5>;
-	type WeightToFee = IdentityFee<Balance>;
-	type LengthToFee = IdentityFee<Balance>;
-	type FeeMultiplierUpdate = ();
 }
 
 impl pallet_sudo::Config for Runtime {
@@ -339,30 +218,6 @@ impl<F: FindAuthor<u32>> FindAuthor<H160> for FindAuthorTruncated<F> {
 		}
 		None
 	}
-}
-
-parameter_types! {
-	pub const ChainId: u64 = 1337;
-	pub BlockGasLimit: U256 = U256::from(u32::max_value());
-	pub PrecompilesValue: FrontierPrecompiles<Runtime> = FrontierPrecompiles::<_>::new();
-}
-
-impl pallet_evm::Config for Runtime {
-	type FeeCalculator = TxHandler;
-	type GasWeightMapping = GafiGasWeightMapping;
-	type BlockHashMapping = pallet_ethereum::EthereumBlockHashMapping<Self>;
-	type CallOrigin = EnsureAddressRoot<AccountId>;
-	type WithdrawOrigin = EnsureAddressNever<AccountId>;
-	type AddressMapping = ProofAddressMapping;
-	type Currency = Balances;
-	type Event = Event;
-	type Runner = pallet_evm::runner::stack::Runner<Self>;
-	type PrecompilesType = FrontierPrecompiles<Self>;
-	type PrecompilesValue = PrecompilesValue;
-	type ChainId = ChainId;
-	type BlockGasLimit = BlockGasLimit;
-	type OnChargeTransaction = GafiEVMCurrencyAdapter<Balances, DealWithFees<Runtime>>;
-	type FindAuthor = FindAuthorTruncated<Aura>;
 }
 
 impl pallet_ethereum::Config for Runtime {
@@ -412,134 +267,6 @@ impl pallet_player::Config for Runtime {
 	type Membership = GafiMembership;
 	type UpfrontPool = UpfrontPool;
 	type StakingPool = StakingPool;
-}
-
-pub const STAKING_BASIC_ID: ID = [0_u8; 32];
-pub const STAKING_MEDIUM_ID: ID = [1_u8; 32];
-pub const STAKING_ADVANCE_ID: ID = [2_u8; 32];
-
-pub const UPFRONT_BASIC_ID: ID = [10_u8; 32];
-pub const UPFRONT_MEDIUM_ID: ID = [11_u8; 32];
-pub const UPFRONT_ADVANCE_ID: ID = [12_u8; 32];
-
-impl SystemDefaultServices for StakingPoolDefaultServices {
-	fn get_default_services() -> [(ID, SystemService); 3] {
-		[
-			(
-				STAKING_BASIC_ID,
-				SystemService::new(
-					STAKING_BASIC_ID,
-					10_u32,
-					Permill::from_percent(30),
-					1000 * unit(GAKI),
-				),
-			),
-			(
-				STAKING_MEDIUM_ID,
-				SystemService::new(
-					STAKING_MEDIUM_ID,
-					10_u32,
-					Permill::from_percent(50),
-					1500 * unit(GAKI),
-				),
-			),
-			(
-				STAKING_ADVANCE_ID,
-				SystemService::new(
-					STAKING_ADVANCE_ID,
-					10_u32,
-					Permill::from_percent(70),
-					2000 * unit(GAKI),
-				),
-			),
-		]
-	}
-}
-
-pub struct UpfrontPoolDefaultServices {}
-
-impl SystemDefaultServices for UpfrontPoolDefaultServices {
-	fn get_default_services() -> [(ID, SystemService); 3] {
-		[
-			(
-				UPFRONT_BASIC_ID,
-				SystemService::new(
-					UPFRONT_BASIC_ID,
-					10_u32,
-					Permill::from_percent(30),
-					5 * unit(GAKI),
-				),
-			),
-			(
-				UPFRONT_MEDIUM_ID,
-				SystemService::new(
-					UPFRONT_MEDIUM_ID,
-					10_u32,
-					Permill::from_percent(50),
-					7 * unit(GAKI),
-				),
-			),
-			(
-				UPFRONT_ADVANCE_ID,
-				SystemService::new(
-					UPFRONT_ADVANCE_ID,
-					10_u32,
-					Permill::from_percent(70),
-					10 * unit(GAKI),
-				),
-			),
-		]
-	}
-}
-
-parameter_types! {
-	pub const MaxPlayerStorage: u32 = 10000;
-}
-
-impl upfront_pool::Config for Runtime {
-	type Event = Event;
-	type Currency = Balances;
-	type WeightInfo = upfront_pool::weights::SubstrateWeight<Runtime>;
-	type MaxPlayerStorage = MaxPlayerStorage;
-	type MasterPool = Pool;
-	type UpfrontServices = UpfrontPoolDefaultServices;
-	type Players = Player;
-}
-
-pub struct StakingPoolDefaultServices {}
-
-impl staking_pool::Config for Runtime {
-	type Event = Event;
-	type Currency = Balances;
-	type WeightInfo = staking_pool::weights::SubstrateWeight<Runtime>;
-	type StakingServices = StakingPoolDefaultServices;
-	type Players = Player;
-}
-
-parameter_types! {
-	pub MinPoolBalance: u128 = 1000 * unit(GAKI);
-	pub MinDiscountPercent: Permill = Permill::from_percent(30);
-	pub MaxDiscountPercent: Permill = Permill::from_percent(70);
-	pub MinTxLimit: u32 = 50;
-	pub MaxTxLimit: u32 = 100;
-	pub MaxPoolOwned: u32 =  10;
-	pub MaxPoolTarget: u32 =  10;
-}
-
-impl sponsored_pool::Config for Runtime {
-	type Event = Event;
-	type Randomness = RandomnessCollectiveFlip;
-	type PoolName = PoolName;
-	type Currency = Balances;
-	type MinPoolBalance = MinPoolBalance;
-	type MinDiscountPercent = MinDiscountPercent;
-	type MaxDiscountPercent = MaxDiscountPercent;
-	type MinTxLimit = MinTxLimit;
-	type MaxTxLimit = MaxTxLimit;
-	type MaxPoolOwned = MaxPoolOwned;
-	type MaxPoolTarget = MaxPoolTarget;
-	type IWhitelist = PalletWhitelist;
-	type WeightInfo = sponsored_pool::weights::SponsoredWeight<Runtime>;
 }
 
 parameter_types! {
@@ -620,7 +347,6 @@ impl game_creator::Config for Runtime {
 	type WeightInfo = game_creator::weights::GameCreatorWeight<Runtime>;
 }
 
-
 impl<LocalCall> frame_system::offchain::CreateSignedTransaction<LocalCall> for Runtime
 where
 	Call: From<LocalCall>,
@@ -630,7 +356,10 @@ where
 		public: <Signature as sp_runtime::traits::Verify>::Signer,
 		account: AccountId,
 		index: Index,
-	) -> Option<(Call, <UncheckedExtrinsic as sp_runtime::traits::Extrinsic>::SignaturePayload)> {
+	) -> Option<(
+		Call,
+		<UncheckedExtrinsic as sp_runtime::traits::Extrinsic>::SignaturePayload,
+	)> {
 		let period = BlockHashCount::get() as u64;
 		let current_block = System::block_number().saturated_into::<u64>().saturating_sub(1);
 		let tip = 0;
@@ -653,7 +382,14 @@ where
 		let signature = raw_payload.using_encoded(|payload| C::sign(payload, public))?;
 		let address = account;
 		let (call, extra, _) = raw_payload.deconstruct();
-		Some((call, (sp_runtime::MultiAddress::Id(address), signature.into(), extra)))
+		Some((
+			call,
+			(
+				sp_runtime::MultiAddress::Id(address),
+				signature.into(),
+				extra,
+			),
+		))
 	}
 }
 
@@ -679,7 +415,7 @@ impl pallet_whitelist::Config for Runtime {
 	type Event = Event;
 	type WhitelistPool = Pool;
 	type SponsoredPool = SponsoredPool;
-	type Currency =  Balances;
+	type Currency = Balances;
 	type WeightInfo = pallet_whitelist::weights::WhitelistWeight<Runtime>;
 	type MaxWhitelistLength = MaxWhitelistLength;
 	type WhitelistFee = WhitelistFee;
@@ -953,40 +689,6 @@ impl pallet_collective::Config<CouncilCollective> for Runtime {
 	type MaxMembers = CouncilMaxMembers;
 	type DefaultVote = pallet_collective::PrimeDefaultVote;
 	type WeightInfo = pallet_collective::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-	pub CandidacyBond: Balance = 100 * centi(GAKI);
-	// 1 storage item created, key size is 32 bytes, value size is 16+16.
-	pub VotingBondBase: Balance = deposit(1, 64, GAKI);
-	// additional data per vote is 32 bytes (account id).
-	pub VotingBondFactor: Balance = deposit(0, 32, GAKI);
-	/// Daily council elections
-	pub TermDuration: BlockNumber = prod_or_fast!(24 * HOURS, 2 * MINUTES, "GAKI_TERM_DURATION");
-	pub const DesiredMembers: u32 = 19;
-	pub const DesiredRunnersUp: u32 = 19;
-	pub const PhragmenElectionPalletId: LockIdentifier = *b"phrelect";
-}
-
-// Make sure that there are no more than MaxMembers members elected via phragmen.
-const_assert!(DesiredMembers::get() <= CouncilMaxMembers::get());
-
-impl pallet_elections_phragmen::Config for Runtime {
-	type Event = Event;
-	type Currency = Balances;
-	type ChangeMembers = Council;
-	type InitializeMembers = Council;
-	type CurrencyToVote = frame_support::traits::U128CurrencyToVote;
-	type CandidacyBond = CandidacyBond;
-	type VotingBondBase = VotingBondBase;
-	type VotingBondFactor = VotingBondFactor;
-	type LoserCandidate = Treasury;
-	type KickedMember = Treasury;
-	type DesiredMembers = DesiredMembers;
-	type DesiredRunnersUp = DesiredRunnersUp;
-	type TermDuration = TermDuration;
-	type PalletId = PhragmenElectionPalletId;
-	type WeightInfo = pallet_elections_phragmen::weights::SubstrateWeight<Runtime>;
 }
 
 impl pallet_hotfix_sufficients::Config for Runtime {
