+++ runtime/devnet/src/pallets/grandpa.rs
@@ -1,0 +1,25 @@
+use frame_support::traits::{ConstU32, KeyOwnerProofSystem};
+use pallet_grandpa::AuthorityId as GrandpaId;
+use sp_runtime::KeyTypeId;
+
+use crate::{Call, Event, Runtime};
+
+impl pallet_grandpa::Config for Runtime {
+	type Event = Event;
+	type Call = Call;
+
+	type KeyOwnerProofSystem = ();
+
+	type KeyOwnerProof =
+		<Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(KeyTypeId, GrandpaId)>>::Proof;
+
+	type KeyOwnerIdentification = <Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(
+		KeyTypeId,
+		GrandpaId,
+	)>>::IdentificationTuple;
+
+	type HandleEquivocation = ();
+
+	type WeightInfo = ();
+	type MaxAuthorities = ConstU32<32>;
+}
