+++ runtime/devnet/src/pallets/mod.rs
@@ -1,0 +1,29 @@
+pub mod system;
+pub use system::*;
+
+pub mod balances;
+pub use balances::*;
+
+pub mod transaction_payment;
+pub use transaction_payment::*;
+
+pub mod evm;
+pub use evm::*;
+
+pub mod timestamp;
+pub use timestamp::*;
+
+pub mod grandpa;
+pub use grandpa::*;
+
+pub mod upfront_pool_config;
+pub use upfront_pool_config::*;
+
+pub mod staking_pool_config;
+pub use staking_pool_config::*;
+
+pub mod sponsored_pool_config;
+pub use sponsored_pool_config::*;
+
+pub mod elections_phragmen;
+pub use elections_phragmen::*;
